fix: update product price type to integer and adjust seeder for price range; modify payment processing to handle amount in cents

+++ src/infrastructure/database/entities/product.orm.entity.ts
@@ -15,7 +15,7 @@ export class ProductORM {
 
   @Column({
     name: 'price',
-    type: 'decimal',
+    type: 'integer',
   })
   price: number;
